// src/services/api/enumEndpoints.ts
export enum EnumEndpoints {
  forgotResetPassword = "/api/password-reset",
  resetPassword = "/api/password-reset/reset",
  registerUser = "/api/auth/register",
  loginUser = "/api/auth/login ",
  logoutUser = "/api/auth/logout",
  updateToken = "/api/auth/token",
  ingredients = "/api/ingredients",
  user = "/api/auth/user",
  orders = "/api/orders",
}